+++ src/App.css
@@ -5,3 +5,7 @@
 
   /* border: 1px solid red; */
 }
+
+html {
+  font-size: 62.5%;
+}
