--- src/component/NowInfo/Payment.jsx
@@ -1,9 +1,0 @@
-function Payment() {
-  return (
-    <>
-      결제 화면입니다.
-    </>
-  )
-}
-
-export default Payment;
