+++ src/pages/ReservationPage/Reservation.jsx
@@ -1,10 +1,12 @@
 import DatePicker from 'react-datepicker';
 import './Reservation.css'
-import { useEffect, useState } from 'react';
 import "react-datepicker/dist/react-datepicker.css";
 import { ko } from 'date-fns/locale';
 import { registerLocale } from 'react-datepicker';
 import { useNavigate } from 'react-router-dom';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faMinus, faPlus } from '@fortawesome/free-solid-svg-icons';
+
 registerLocale('ko', ko);
 
 function Reservation({rsvnInfo, setRsvnInfo, checkInDate, checkOutDate, totalGuestCount}) {
@@ -25,88 +27,103 @@ function Reservation({rsvnInfo, setRsvnInfo, checkInDate, checkOutDate, totalGue
 
   return (
     <div className="reservation">
-      <div className="reservationTitle">
-        <h3>예약</h3>
-        <h2>날짜 및 인원 선택</h2>
-      </div>
-
-      {/* 투숙 날짜 선택 */}
-      <div className="dateSelector">
-          <DatePicker 
-            showIcon
-            selected={startDate}
-            onChange={onChange}
-            startDate={startDate}
-            endDate={endDate}
-            minDate={new Date()}
-            selectsRange
-            inline
-            locale="ko"
-            dateFormat="yyyy/MM/dd"
-          />
-        <div className="checkInOut">
-          <p>체크인</p>
-          <div className="checkInOutDate">
-            {checkInDate}
-          </div>
+      <div className="rsvnWrap1">
+        <div className="rsvnTitle">
+          <h3>예약</h3>
+          <h2>날짜 및 인원 선택</h2>
         </div>
-        <div className="checkInOut">
-          <p>체크아웃</p>
-          <div className="checkInOutDate">
-            {checkOutDate}
-          </div>
-        </div>
-      </div>
 
-      {/* 투숙 인원 선택 */}
-      <div className="selectBox">
-        <div className="countSelect">
-          <p>성인</p>
-          <div className="countSelectBtn">
-            <button onClick={()=>{
-              if(adultCount>1) {
-                updateInfo('adultCount', adultCount-1);
-              }
-            }}>-</button>
-            <p>{adultCount}</p>
-            <button onClick={()=>{
-              if(totalGuestCount<3 && adultCount<3) {
-                updateInfo('adultCount', adultCount+1);
-              }
-            }}>+</button>
+        {/* 투숙 날짜 선택 */}
+        <div className="dateSelector">
+          <div className="calendarBox">
+            <DatePicker 
+              showIcon
+              selected={startDate}
+              onChange={onChange}
+              startDate={startDate}
+              endDate={endDate}
+              minDate={new Date()}
+              monthsShown={2}
+              selectsRange
+              inline
+              locale="ko"
+              dateFormat="yyyy/MM/dd"
+            />
           </div>
         </div>
 
-        <div className="countSelect">
-          <p>어린이</p>
-          <div className="countSelectBtn">
-            <button onClick={()=>{
-              if(childCount>0) {
-                updateInfo('childCount', childCount-1);
-              }
-            }}>-</button>
-            <p>{childCount}</p>
-            <button onClick={()=>{
-              if(totalGuestCount<3 && childCount<2) {
-                updateInfo('childCount', childCount+1);
-              }
-            }}>+</button>
+        <div className="selectBox">
+          <div className='selectItem'>
+            <p>체크인</p>
+            <div className='selectContent'>
+              <div className="checkInOutDate">
+                {checkInDate}
+              </div>
+            </div>
+          </div>
+          <div className='selectItem'>
+            <p>체크아웃</p>
+            <div className='selectContent'>
+              <div className="checkInOutDate">
+                {checkOutDate}
+              </div>
+            </div>
           </div>
         </div>
-      </div>
 
-      <div className="nextBtn" onClick={()=>{
-        if(checkOutDate == '') {
-          alert('체크아웃 날짜를 선택하세요');
-        } else if(checkOutDate == checkInDate) {
-          alert('최소 1박 이상으로 선택하세요')
-        } else {
-          navigate('/reservation/room');
-        }
-      }}>
-        다음 &gt;
+        {/* 투숙 인원 선택 */}
+        <div className="selectBox">
+          <div className='selectItem'>
+            <p>인원</p>
+            <div className='selectContent'>
+              <div className="countSelect">
+                <p>성인</p>
+                <div className="countSelectBtn">
+                  <button onClick={()=>{
+                    if(adultCount>1) {
+                      updateInfo('adultCount', adultCount-1);
+                    }
+                  }}><FontAwesomeIcon icon={faMinus} /></button>
+                  <p>{adultCount}</p>
+                  <button onClick={()=>{
+                    if(totalGuestCount<3 && adultCount<3) {
+                      updateInfo('adultCount', adultCount+1);
+                    }
+                  }}><FontAwesomeIcon icon={faPlus} /></button>
+                </div>
+              </div>
+              <div className="countSelect">
+                <p>어린이</p>
+                <div className="countSelectBtn">
+                  <button onClick={()=>{
+                    if(childCount>0) {
+                      updateInfo('childCount', childCount-1);
+                    }
+                  }}><FontAwesomeIcon icon={faMinus} /></button>
+                  <p>{childCount}</p>
+                  <button onClick={()=>{
+                    if(totalGuestCount<3 && childCount<2) {
+                      updateInfo('childCount', childCount+1);
+                    }
+                  }}><FontAwesomeIcon icon={faPlus} /></button>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        <div className="nextBtn" onClick={()=>{
+          if(checkOutDate == '') {
+            alert('체크아웃 날짜를 선택하세요');
+          } else if(checkOutDate == checkInDate) {
+            alert('최소 1박 이상으로 선택하세요')
+          } else {
+            navigate('/reservation/room');
+          }
+        }}>
+          다음 &gt;
+        </div>
       </div>
-      
     </div>
   )
 }
